+++ src/main/java/com/neovisionaries/ws/client/PeriodicalFrameSender.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015-2016 Neo Visionaries Inc.
+ * Copyright (C) 2015-2018 Neo Visionaries Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ import java.util.TimerTask;
 abstract class PeriodicalFrameSender
 {
     private final WebSocket mWebSocket;
-    private final String mTimerName;
+    private String mTimerName;
     private Timer mTimer;
     private boolean mScheduled;
     private long mInterval;
@@ -95,13 +95,19 @@ abstract class PeriodicalFrameSender
         {
             if (mTimer == null)
             {
-                mTimer = new Timer(mTimerName);
+                if (mTimerName == null)
+                {
+                    mTimer = new Timer();
+                }
+                else
+                {
+                    mTimer = new Timer(mTimerName);
+                }
             }
 
             if (mScheduled == false)
             {
-                mScheduled = true;
-                mTimer.schedule(new Task(), interval);
+                mScheduled = schedule(mTimer, new Task(), interval);
             }
         }
     }
@@ -121,6 +127,21 @@ abstract class PeriodicalFrameSender
         synchronized (this)
         {
             mGenerator = generator;
+        }
+    }
+
+
+    public String getTimerName()
+    {
+        return mTimerName;
+    }
+
+
+    public void setTimerName(String timerName)
+    {
+        synchronized (this)
+        {
+            mTimerName = timerName;
         }
     }
 
@@ -151,7 +172,7 @@ abstract class PeriodicalFrameSender
             mWebSocket.sendFrame(createFrame());
 
             // Schedule a new task.
-            mTimer.schedule(new Task(), mInterval);
+            mScheduled = schedule(mTimer, new Task(), mInterval);
         }
     }
 
@@ -182,6 +203,38 @@ abstract class PeriodicalFrameSender
         {
             // Empty payload.
             return null;
+        }
+    }
+
+
+    private static boolean schedule(Timer timer, Task task, long interval)
+    {
+        try
+        {
+            // Schedule the task.
+            timer.schedule(task, interval);
+
+            // Successfully scheduled the task.
+            return true;
+        }
+        catch (RuntimeException e)
+        {
+            // Failed to schedule the task. Probably, the exception is
+            // an IllegalStateException which is raised due to one of
+            // the following reasons (according to the Javadoc):
+            //
+            //   (1) if task was already scheduled or cancelled,
+            //   (2) timer was cancelled, or
+            //   (3) timer thread terminated.
+            //
+            // Because a new task is created every time this method is
+            // called and there is no code to call TimerTask.cancel(),
+            // (1) cannot be a reason.
+            //
+            // In either case of (2) and (3), we don't have to retry to
+            // schedule the task, because the timer that is expected to
+            // host the task will stop or has stopped anyway.
+            return false;
         }
     }